requirements.txt
streamlit
groq
pdfplumber
pillow
pytesseract
